requirements.txt
stanford_openie
nltk
spacy-dbpedia-spotlight
progress
SPARQLWrapper
